Recommended way to be installed in GNU/Linux implies using sudo.
The issue is known to be happening on Debian/Ubuntu systems because they do not add `~/.local/bin` to the path as said in <https://github.com/pypa/pip/issues/3813>.
It looks like it is being solved on Debian testing by automatically add this. Pip devs do not seem to develop a given fix on that.